requirements.txt
flask==3.1.0
flask-cors==5.0.0
openai==1.82.0
python-dotenv==1.0.1
gunicorn==23.0.0